server/server/app/worker/requirements.txt
aiofiles==24.1.0
aiohttp==3.10.3
alembic==1.13.2
asyncpg==0.29.0
bamtools
biopython==1.84
bracken
bwa
celery[redis]==5.4.0
fastapi[all]==0.112.0
fastp==0.23.2
fastqc==0.11.9
flower==2.0.1
kraken2 
multiqc==1.13
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-multipart==0.0.9
qualimap 
quast==5.3.0
r-base
redis-py==5.0.9
samtools 
spades==3.15.5
sqlalchemy==2.0.25
uvicorn[standard]==0.37.0